changed mappedto name to display name universally, renamed and moved some endpoints
