+++ api/routes/club/club.ts
@@ -4,9 +4,106 @@ import auth from '../../middleware/auth';
 import jwt from 'jsonwebtoken';
 import { oAuth2Client } from '../auth/google';
 import { google } from 'googleapis';
+import { sheets_v4 as sheets } from 'googleapis';
 
 export const router = Router();
 const prisma = new PrismaClient();
+
+router.get(
+  '/size/:organisationId',
+  auth,
+  async (req: Request, res: Response) => {
+    try {
+      const organisationId = Number.parseInt(req.params['organisationId']);
+
+      const pages = await prisma.page.findMany({
+        where: {
+          organisationId: organisationId,
+        },
+        select: {
+          id: true,
+          sheetId: true,
+          sheetTabId: true,
+          identificationColumns: {
+            select: {
+              originalName: true,
+            },
+          },
+        },
+      });
+
+      if (!pages) {
+        return res
+          .status(500)
+          .send(`No Pages found for organisation ${organisationId}`);
+      }
+
+      const memberCountByPageId: IMemberCountByPageId = {};
+
+      for (let i = 0; i < pages.length; i++) {
+        const page = pages[i];
+
+        const metadataResponse = await sheets.spreadsheets.get({
+          spreadsheetId: page.sheetId,
+        });
+
+        if (!metadataResponse.data.sheets) {
+          return res
+            .status(400)
+            .send(JSON.stringify('spreadsheet has no sheets'));
+        }
+
+        const sheet = metadataResponse.data.sheets.find(
+          (sheet) => sheet.properties?.sheetId === Number(page.sheetTabId)
+        );
+
+        if (!sheet) {
+          return res.status(400).send(JSON.stringify('sheet not found'));
+        }
+
+        const sheetName = sheet.properties?.title;
+
+        const range = `${sheetName}!A1:Z`;
+        const {
+          data: { values },
+        } = await sheets.spreadsheets.values.get({
+          spreadsheetId: page.sheetId,
+          range: range,
+        });
+
+        const columnNames = [];
+
+        for (let j = 0; j < page.identificationColumns.length; j++) {
+          columnNames.push(page.identificationColumns[j].originalName);
+        }
+
+        if (values) {
+          const headers = values[0];
+
+          for (let k = 0; k < headers.length; k++) {
+            if (columnNames.includes(headers[k])) {
+              const columnData = values.slice(1).map((row) => row[k]);
+
+              const nonEmptyRowCount = columnData.filter(
+                (cellValue) => cellValue !== '' && cellValue !== undefined
+              ).length;
+
+              memberCountByPageId[page.id] = {
+                UniqueColumnName: headers[k],
+                totalMembers: nonEmptyRowCount,
+              };
+            }
+          }
+        }
+      }
+
+      return res.status(200).send(memberCountByPageId);
+    } catch (err) {
+      console.error(err);
+      return res.status(500).send('failed to retrieve list of admins');
+    }
+  }
+);
 
 router.post('/create', auth, async (req, res) => {
   const { clubName, clubAcronym } = req.body;
@@ -111,14 +208,14 @@ router.get(
 
       const organisation = await prisma.organisation.findUnique({
         where: {
-          name: organisationName
-        }});
+          name: organisationName,
+        },
+      });
 
-        if (!organisation) return res.status(400).send('invalid club name');
-        const organisationId = organisation.id;
-      return res.status(200).send({organisationId});
-    }
-      catch (err) {
+      if (!organisation) return res.status(400).send('invalid club name');
+      const organisationId = organisation.id;
+      return res.status(200).send({ organisationId });
+    } catch (err) {
       console.error(err);
       return res.status(500).send('failed to get organisation id');
     }
